Updated the code pointed out and quit using JpaRepository class
