--- backend/src/main/java/com/example/foa/model/dto/UserDTO.java
@@ -1,24 +1,0 @@
-package com.example.foa.model.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDTO {
-    private Long id;
-
-    @NotBlank(message = "名前は必須です")
-    @Size(min = 1, max = 100, message = "名前は1-100文字で入力してください")
-    private String name;
-
-    @NotBlank(message = "メールアドレスは必須です")
-    @Email(message = "有効なメールアドレスを入力してください")
-    private String email;
-}
