+++ backend/src/main/java/dev/foa/backend/Application.java
@@ -6,8 +6,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 @EnableJpaAuditing
-public class FoaBackendApp {
+public class Application {
     public static void main(String[] args) {
-        SpringApplication.run(FoaBackendApp.class, args);
+
+        SpringApplication.run(Application.class, args);
+
     }
 }
