+++ backend/src/main/java/dev/foa/backend/controller/RootController.java
@@ -1,6 +1,4 @@
-package com.example.foa.controller;
-
-import java.util.Map;
+package dev.foa.backend.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -8,11 +6,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.Map;
+
 @RestController
 @RequestMapping("")
 @CrossOrigin(origins = "http://localhost:3000")
 public class RootController {
-    @GetMapping
+    @GetMapping({ "", "/" })
     public ResponseEntity<Map<String, String>> root() {
         return ResponseEntity.ok(Map.of("status", "ok"));
     }
