+++ backend/src/main/java/dev/foa/backend/exception/UserAlreadyExistsException.java
