+++ backend/src/main/java/dev/foa/backend/exception/UserExceptionHandler.java
@@ -1,0 +1,4 @@
+package dev.foa.backend.exception;
+
+public class UserExceptionHandler {
+}
