+++ backend/src/main/java/dev/foa/backend/exception/UserNotFoundException.java
