+++ backend/src/main/java/dev/foa/backend/util/UserJsonDataLoaderUtil.java
@@ -1,0 +1,4 @@
+package dev.foa.backend.util;
+
+public class UserJsonDataLoaderUtil {
+}
