--- backend/src/test/java/com/example/foa/FoaBackendAppTests.java
@@ -1,11 +1,0 @@
-package com.example.foa;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FoaBackendAppTests {
-	public static void main(String[] args) {
-		SpringApplication.run(FoaBackendAppTests.class, args);
-	}
-}
